Chat input accessory view has been created.
