+++ Tinder/Model/Message.swift
@@ -1,0 +1,11 @@
+//
+//  Message.swift
+//  Tinder
+//
+//  Created by Turan Çabuk on 28.02.2024.
+//
+
+struct Message {
+    let text: String
+    let isFromCurrentUser: Bool
+}
